# Support ignoring particular files in `@path`.