@fonsp, @dralletje this might be helpful for those source map files that were getting included in Pluto's relocatable folders. Just ignore the source map files and we might improve load time, at the cost of not having source maps in relocated images with Pluto. Let me know if it speeds things up.