@fonsp 2022-02-23
This would be great to have for another project, where I want to ignore the `node_modules` folder!