@fonsp 2022-02-23
Code coverage should be fixed by https://github.com/JuliaPackaging/RelocatableFolders.jl/pull/14