@fonsp 2022-02-24
I tried this on a pretty big `frontend` folder with original source, `node_modules` and build output. We ignore everything except build output.

<details>
<summary>Benchmarks</summary>




# Empty
Code:
```julia
module Example
using RelocatableFolders
end
```

Output:
```
julia> @time Pkg.precompile()
Precompiling project...
  1 dependency successfully precompiled in 2 seconds (305 already precompiled)
  3.640369 seconds (6.86 M allocations: 449.645 MiB, 5.48% gc time, 12.38% compilation time)

julia> @time using Example
  0.011561 seconds (10.19 k allocations: 493.312 KiB, 31.66% compilation time)

```

# Before
Code:
```julia
module Example
using RelocatableFolders

const dir = @path joinpath(@__DIR__, "../frontend")
readdir(dir)

end
```

Output:
```
julia> @time Pkg.precompile()
Precompiling project...
  1 dependency successfully precompiled in 15 seconds (305 already precompiled)
 15.866306 seconds (7.01 M allocations: 479.562 MiB, 1.46% gc time, 2.91% compilation time)

julia> @time using CatalystGUI
  0.536946 seconds (268.63 k allocations: 303.097 MiB, 0.54% compilation time)


```
# After
Code:
```julia
module Example
using RelocatableFolders

const dir = @path(joinpath(@__DIR__, "../frontend"), !occursin("build"))
readdir(dir)

end
```

Output:
```
julia> @time Pkg.precompile()
Precompiling project...
  1 dependency successfully precompiled in 4 seconds (305 already precompiled)
  5.807133 seconds (6.97 M allocations: 478.807 MiB, 3.72% gc time, 8.46% compilation time)

julia> @time using CatalystGUI
  0.014067 seconds (10.34 k allocations: 507.977 KiB, 25.75% compilation time)

```

</details>

It's working!

(Writing this benchmark made me realise that I don't need this feature for what I want, I can just use `@path` directly on the folder containing build output. 🤦)