@MichaelHatherly 2022-02-24
> (Writing this benchmark made me realise that I don't need this feature for what I want, I can just use `@path` directly on the folder containing build output. :facepalm:)

:rofl: ah well, it'll be useful in other circumstances anyway. Thanks for checking.