@fonsp 2022-02-25
> One issue with how it currently checks for inclusion of files is to iterate over all files, regardless whether the directory would be ignored or not and ignore on a file-basis only. Probably not an issue in most cases, though with a `node_modules/` it may slow down precomp of the module what's using `@path`, though package load time shouldn't be affected I would think.

You might be able to take some ideas from https://github.com/shashi/FileTrees.jl and https://github.com/vtjnash/Glob.jl#readme if this ever comes up again